requirements.txt
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
mysqlclient==2.2.4
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.3
email-validator==2.2.0
beautifulsoup4==4.12.3
cryptography==44.0.1
django-redis==5.4.0
gunicorn==22.0.0
celery==5.4.0
